Python_Gateway/requirements.txt
keras
tensorflow
Pillow
numpy
opencv-python
pyserial
adafruit-io
